Replace session-based authentication with JWT
This commit removes session-based authentication middleware and replaces it with JWT authentication and bearer token. In addition to this, the Bcrypt library has been introduced to securely hash passwords prior to saving them in the database. Unnecessary packages related to express-session and mongoose-field-encryption have been removed. This commit also updates all the routes to use the new authentication method.