Refactor import statements for SsmAnsible and SettingsKeys
The change primarily involves refining the import statements to streamline the code base. The refactor includes importing SettingsKeys in the startup module and SsmAnsible across multiple modules and changing the encryption logic method in DockerAPIHelper.test. Other changes include minor tweaks in the async/await functions and the removal of an unused import in the shell module.