Update package dependencies
The package dependencies for both the server and client side have been updated. The libraries "mongoose", "@aws-sdk/client-ecr", and "eslint" for the server, and "antd" and "eslint" for the client, have been updated to newer versions. These updates contain bug fixes and improvements which help keep our application secure and efficient.