Refactor ExtraVarsTransformer and update tests
The ExtraVarsTransformer in the Ansible utility has been refactored for better handling of empty entries. The corresponding tests have also been updated. Now, if the input is empty, it will automatically return an empty object rather than throw an error. For missing value field in extraVar, it will return an empty string instead of 'undefined'.