Add default local repository for playbooks on user creation
Added an automatic default local user repository for playbooks when a new user is created. This was done by modifying startup logic to include a default user creation function, and updating the UserRepo to support this change. The SCHEME_VERSION in settings has also been updated to reflect the change.